refactor: Move code from `executable_command.go` and `directory_module.go` to `contract.go`
