+++ contract_test.go
